Accueil done but img choice and accordion desktop need help
