+++ src/components/Accueil/modes.module.css
@@ -1,0 +1,79 @@
+/*
+Charte graphique - couleurs
+#D6D6D6 GRIS CLAIRE
+#FFEE32 JAUNE CLAIR
+#FFD100 JAUNE
+#202020 NOIR
+#333533 GRIS FONCE
+
+
+ font-family: 'Chakra Petch', sans-serif;
+*/
+@import url('https://fonts.googleapis.com/css2?family=Chakra+Petch:wght@300&family=Kumar+One&display=swap');
+html body {
+  padding: 0;
+  margin: 0;
+  background-color: #d6d6d6;
+}
+.menuModesContainer {
+  background-color: #d6d6d6;
+}
+.menuHeaderDesk {
+  display: none;
+}
+button {
+  width: 100%;
+  font-family: 'Kumar', sans-serif;
+  color: #ffd100;
+  background-color: #333533;
+  text-align: center;
+}
+
+img {
+  max-width: 100%;
+  position: relative;
+}
+
+.innerMenuDesc {
+  position: absolute;
+  width: 90%;
+  text-align: center;
+  color: #ffd100;
+  background-color: #333533;
+  opacity: 0.8;
+  padding: 0.5em;
+  font-size: 1.2em;
+}
+
+.menuContent {
+  display: flex;
+  box-align: center;
+  align-items: flex-end;
+  justify-content: center;
+}
+
+@media screen and (min-width: 768px) {
+  .menuHeaderDesk {
+    display: block;
+    font-style: italic;
+    font-size: x-large;
+    text-align: center;
+  }
+  .menuModesContainer {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+  }
+
+  .menuChallenge {
+    width: 33%;
+  }
+
+  .menuContent {
+    width: 100%;
+    height: 50vh;
+  }
+  .innerMenuDesc {
+    max-width: 30%;
+  }
+}
