--- src/components/MenuModes.css
@@ -1,55 +1,0 @@
-/*
-Charte graphique - couleurs
-#D6D6D6 GRIS CLAIRE
-#FFEE32 JAUNE CLAIR
-#FFD100 JAUNE
-#202020 NOIR
-#333533 GRIS FONCE
-
-
- font-family: 'Chakra Petch', sans-serif;
-*/
-@import url('https://fonts.googleapis.com/css2?family=Chakra+Petch:wght@300&family=Kumar+One&display=swap');
-html body {
-  padding: 0;
-  margin: 0;
-}
-
-.container {
-  font-family: 'Chakra petch', sans-serif;
-  color: #ffd100;
-  background-color: #333533;
-  text-align: center;
-}
-
-.menuMode-container {
-  display: flex;
-  flex-direction: column;
-}
-
-.menu-title,
-.inner-menu-title {
-  font-family: 'Kumar one', sans-serif;
-  color: #ffd100;
-  background-color: #333533;
-  text-align: center;
-}
-
-.inner-menu-title {
-  font-family: 'Kumar one', sans-serif;
-  color: #ffd100;
-  background-color: #333533;
-}
-
-.inner-menu-desc {
-  font-family: chakra petch;
-  font-weight: bold;
-  font-style: italic;
-  color: #ffd100;
-  background-color: #333533;
-  opacity: 0.7;
-  text-align: end;
-}
-
-@media screen and (min-width: 768px) {
-}
